moved the Indicators to Operators
