+++ include/shark/Algorithms/DirectSearch/SMS-EMOA.h
@@ -38,7 +38,7 @@
 
 // MOO specific stuff
 #include <shark/Algorithms/DirectSearch/Individual.h>
-#include <shark/Algorithms/DirectSearch/Indicators/HypervolumeIndicator.h>
+#include <shark/Algorithms/DirectSearch/Operators/Indicators/HypervolumeIndicator.h>
 #include <shark/Algorithms/DirectSearch/Operators/Selection/TournamentSelection.h>
 #include <shark/Algorithms/DirectSearch/Operators/Selection/IndicatorBasedSelection.h>
 #include <shark/Algorithms/DirectSearch/Operators/Recombination/SimulatedBinaryCrossover.h>
